+++ backend/canisters/registry/impl/src/updates/add_message_filter.rs
@@ -4,7 +4,7 @@ use canister_tracing_macros::trace;
 use registry_canister::add_message_filter::{Response::*, *};
 use user_index_canister_c2c_client::{lookup_user, LookupUserError};
 
-#[update(candid = true, msgpack = true)]
+#[update(msgpack = true)]
 #[trace]
 async fn add_message_filter(args: Args) -> Response {
     if args.regex.len() > 1000 {