+++ backend/canisters/notifications_index/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Removed
+
+- Remove deprecated candid endpoints ([#6396](https://github.com/open-chat-labs/open-chat/pull/6396))
+
 ## [[2.0.1333](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1333-notifications_index)] - 2024-09-06
 
 ### Added